requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
atlastk==0.13.5
ciso8601==2.3.3
click==8.3.1
fastapi==0.135.1
h11==0.16.0
httptools==0.7.1
idna==3.11
iso8601==2.1.0
orjson==3.11.7
pydantic==2.12.5
pydantic_core==2.41.5
pypika-tortoise==0.6.5
python-dotenv==1.2.2
PyYAML==6.0.3
SQLAlchemy==2.0.48
starlette==1.0.0
tortoise-orm==1.1.7
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
